+++ src/core/sync/index.ts
@@ -1,2 +1,3 @@
 export * from "./LocalOnlySyncProvider";
+export * from "./optionalSyncConsent";
 export * from "./types";